Add rule integration tests for all other supported languages
