+++ pmd-dist/src/test/java/net/sourceforge/pmd/it/BinaryDistributionIT.java
@@ -9,47 +9,17 @@ import static org.junit.Assert.fail;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Path;
 import java.util.Enumeration;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipFile;
 
-import org.apache.commons.io.FileUtils;
-import org.junit.AfterClass;
-import org.junit.BeforeClass;
 import org.junit.Test;
 
 import net.sourceforge.pmd.PMDVersion;
 
-public class BinaryDistributionIT {
-
-    private static File getBinaryDistribution() {
-        return new File(".", "target/pmd-bin-" + PMDVersion.VERSION + ".zip");
-    }
-
-    /**
-     * The temporary directory, to which the binary distribution will be extracted.
-     * It will be deleted again after the test.
-     */
-    private static Path tempDir;
-
-    @BeforeClass
-    public static void setupTempDirectory() throws Exception {
-        tempDir = Files.createTempDirectory("pmd-it-test-");
-        if (getBinaryDistribution().exists()) {
-            ZipFileExtractor.extractZipFile(getBinaryDistribution().toPath(), tempDir);
-        }
-    }
-
-    @AfterClass
-    public static void cleanupTempDirectory() throws IOException {
-        if (tempDir != null && tempDir.toFile().exists()) {
-            FileUtils.forceDelete(tempDir.toFile());
-        }
-    }
+public class BinaryDistributionIT extends AbstractBinaryDistributionTest {
 
     @Test
     public void testFileExistence() {
@@ -102,32 +72,6 @@ public class BinaryDistributionIT {
 
         result = PMDExecutor.runPMDRules(tempDir, srcDir, "rulesets/java/quickstart.xml");
         result.assertExecutionResult(4, "");
-    }
-
-    @Test
-    public void testAllJavaRules() throws Exception {
-        String srcDir = new File(".", "src/test/resources/sample-source/java/").getAbsolutePath();
-
-        ExecutionResult result = PMDExecutor.runPMDRules(tempDir, srcDir, "src/test/resources/rulesets/all-java.xml");
-        assertDefaultExecutionResult(result);
-    }
-
-    @Test
-    public void testAllApexRules() throws Exception {
-        String srcDir = new File(".", "src/test/resources/sample-source/apex/").getAbsolutePath();
-
-        ExecutionResult result = PMDExecutor.runPMDRules(tempDir, srcDir, "src/test/resources/rulesets/all-apex.xml");
-        assertDefaultExecutionResult(result);
-    }
-
-    private static void assertDefaultExecutionResult(ExecutionResult result) {
-        result.assertExecutionResult(4, "");
-
-        result.assertNoError("Exception applying rule");
-        result.assertNoError("Ruleset not found");
-        result.assertNoError("Use of deprecated attribute");
-        result.assertNoErrorInReport("Error while processing");
-        result.assertNoErrorInReport("Error while parsing");
     }
 
     @Test
