+++ pmd-dist/src/test/resources/rulesets/all-javascript.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+
+<ruleset name="All JavaScript Rules" 
+    xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 https://pmd.sourceforge.io/ruleset_2_0_0.xsd">
+    <description>Every JavaScript Rule in PMD</description>
+
+    <rule ref="category/ecmascript/bestpractices.xml" />
+    <rule ref="category/ecmascript/codestyle.xml" />
+    <rule ref="category/ecmascript/design.xml" />
+    <rule ref="category/ecmascript/documentation.xml" />
+    <rule ref="category/ecmascript/errorprone.xml" />
+    <rule ref="category/ecmascript/multithreading.xml" />
+    <rule ref="category/ecmascript/performance.xml" />
+    <rule ref="category/ecmascript/security.xml" />
+
+</ruleset>
