+++ pmd-dist/src/test/resources/rulesets/all-visualforce.xml
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+
+<ruleset name="All Visualforce Rules" 
+    xmlns="http://pmd.sourceforge.net/ruleset/2.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://pmd.sourceforge.net/ruleset/2.0.0 https://pmd.sourceforge.io/ruleset_2_0_0.xsd">
+    <description>Every Visualforce Rule in PMD</description>
+
+    <rule ref="category/vf/bestpractices.xml" />
+    <rule ref="category/vf/codestyle.xml" />
+    <rule ref="category/vf/design.xml" />
+    <rule ref="category/vf/documentation.xml" />
+    <rule ref="category/vf/errorprone.xml" />
+    <rule ref="category/vf/multithreading.xml" />
+    <rule ref="category/vf/performance.xml" />
+    <rule ref="category/vf/security.xml" />
+
+</ruleset>
