+++ pmd-dist/src/test/resources/sample-source/javascript/SampleCode.js
@@ -1,0 +1,7 @@
+(function() {
+    if (true) {
+       x=2;
+    } else {
+       x=4;
+    }
+})();
