+++ pmd-dist/src/test/resources/sample-source/jsp/SampleCode.jsp
@@ -1,0 +1,12 @@
+<%@ page errorPage="error.jsp" %>
+<%@ page import="com.foo.AClass"%>
+<%@ page import="com.foo.MyClass"%>
+
+<html>
+  <body>
+    <b>
+      <img src="<%=Some.get()%>/foo">xx</img>
+      text
+    </b>
+  </body>
+</html>
