+++ pmd-dist/src/test/resources/sample-source/pom/pom.xml.pom
@@ -1,0 +1,17 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <groupId>net.sourceforge.pmd</groupId>
+    <artifactId>xml-pom</artifactId>
+    <version>1.0.0-SNAPSHOT</version>
+
+  <dependencyManagement>
+    <dependency>
+      <groupId>org.jboss.arquillian</groupId>
+      <artifactId>arquillian-bom</artifactId>
+      <version>${arquillian.version}</version>
+      <type>bom</type>
+      <scope>import</scope>
+    </dependency>
+  </dependencyManagement>
+
+</project>
