+++ pmd-dist/src/test/resources/sample-source/visualforce/SampleCode.page
@@ -1,0 +1,1 @@
+<apex:page controller="AcRestActionsController" action="{!csrfInitMethod}" >
