+++ pmd-dist/src/test/resources/sample-source/xsl/samplecode.xslt
@@ -1,0 +1,4 @@
+<xsl:stylesheet version="1.0"
+                xmlns:xsl="http://www.w3.org/1999/XSL/Transform">
+  <xsl:variable name="var" select="concat('abc',concat('cd','ef'))"/>
+</xsl:stylesheet>
